Add an even more prominent warning to wx/msw/ole/comimpl.h
The stuff here is either completely obsolete (wxAutoLong) or broken, so
there is really no good reason to use it any more and it should be
arguably removed completely, but this would require modifying all the
existing code in wxWidgets itself which still uses it.

The warning added in 0caddb4472 (Document important COM implementation
macro limitation, 2022-05-27) was not clear enough, so make it as
obvious as possible that it shouldn't be used now.