wxQt: Override wxDC::{Get,Set}LayoutDirection()
Notice that we do not call QPainter::setLayoutDirection(Qt::RightToLeft) in an RTL
layout, as this only affects text drawing. Also, calling it on a custom-mirrored
DC (i.e., with scale(-1, 1)) does not work as expected. Instead, we explicitly store
the layout direction in wxDC and never call QPainter::setLayoutDirection().